requirements.txt
h5py==2.10.0
fire==0.4.0
scikit_allel==1.3.2
matplotlib==3.3.2
zarr==2.6.1
numcodecs==0.7.3
pandas==1.1.3
tqdm==4.50.2
Bio==0.4.1
